Extracted pension pot filter methods into module
